Added new event handling system.
